+++ es/index.js
@@ -1,4 +1,5 @@
 import { isFunction } from 'lodash';
+import { readable } from 'is-stream';
 import { mergeObjects } from '@lykmapipo/common';
 import { getString } from '@lykmapipo/env';
 import { Router } from '@lykmapipo/express-common';
@@ -111,6 +112,80 @@ const schemaFor = optns => {
 
   // return http getSchema handler
   return httpGetSchema;
+};
+
+/**
+ * @function downloadFor
+ * @name downloadFor
+ * @description Create http get handler for downloading of a given
+ * service options
+ * @param {Object} optns valid downloadFor options
+ * @param {Function} optns.download valid service to to invoke when
+ * downloading. It must return `readStream` which is `stream.Readable` and
+ * `fileName` which is `String`.
+ * @return {Function} valid express middleware to handle downloading request
+ * @author lally elias <lallyelias87@mail.com>
+ * @license MIT
+ * @since 0.3.0
+ * @version 0.1.0
+ * @static
+ * @public
+ * @example
+ *
+ * const { createReadStream } = require('fs');
+ * const { app, downloadFor } = require('@lykmapipo/express-rest-actions');
+ *
+ * const download = (query, done) => {
+ *    const fileName = 'avatar.png';
+ *    const readStream = createReadStream('./avatar.png');
+ *    return done(null, { fileName, readStream });
+ * };
+ *
+ * app.get('/v1/files/avatar', downloadFor({ download }));
+ *
+ */
+const downloadFor = optns => {
+  // ensure options
+  const options = mergeObjects(optns);
+  const { download: doDownload, status = 200 } = options;
+
+  // create http handler to download
+  const httpDownload = (request, response, next) => {
+    // ensure service download and fileName provider
+    if (!isFunction(doDownload)) {
+      return response.methodNotAllowed();
+    }
+
+    // obtain mquery options
+    const query = mergeObjects(request.mquery);
+
+    // handle request
+    const afterHttpDownload = (error, results) => {
+      // handle error
+      if (error) {
+        return next(error);
+      }
+
+      // obtain download file name and stream
+      const { fileName, readStream } = results;
+
+      // ensure read stream
+      if (!readable(readStream)) {
+        return response.methodNotAllowed();
+      }
+
+      // handle success downloading
+      response.attachment(fileName);
+      response.status(status);
+      return readStream.pipe(response);
+    };
+
+    // invoke service download
+    return doDownload(query, afterHttpDownload);
+  };
+
+  // return http download handler
+  return httpDownload;
 };
 
 /**
@@ -397,6 +472,9 @@ const deleteFor = optns => {
  * @param {Function} optns.get valid service function to invoke when get
  * @param {Function} [optns.getSchema] valid service function to invoke when
  * get schema
+ * @param {Function} [optns.export] valid service function to invoke when get
+ * exports. It must return `readStream` which is `stream.Readable` and
+ * `fileName` which is `String`.
  * @param {Function} optns.getById valid service function to invoke when getById
  * @param {Function} optns.post valid service function to invoke when post
  * @param {Function} optns.patch valid service function to invoke when patch
@@ -424,10 +502,14 @@ const routerFor = optns => {
   const defaults = { version: getString('API_VERSION', '1.0.0'), soft: false };
   const options = mergeObjects(defaults, optns);
 
+  // normalize download and export handler
+  options.download = options.download || options.export;
+
   // create paths
   const { pathSingle = `/${options.resource}/:id` } = options;
   const { pathList = `/${options.resource}` } = options;
   const { pathSchema = `/${options.resource}/schema` } = options;
+  const { pathExport = `/${options.resource}/export` } = options;
 
   // create versioned router
   const router = new Router(options);
@@ -435,6 +517,7 @@ const routerFor = optns => {
   // bind http action handlers
   router.get(pathList, getFor(options));
   router.get(pathSchema, schemaFor(options));
+  router.get(pathExport, downloadFor(options));
   router.get(pathSingle, getByIdFor(options));
   router.post(pathList, postFor(options));
   router.patch(pathSingle, patchFor(options));
@@ -445,4 +528,4 @@ const routerFor = optns => {
   return router;
 };
 
-export { deleteFor, getByIdFor, getFor, patchFor, postFor, putFor, routerFor, schemaFor };
+export { deleteFor, downloadFor, getByIdFor, getFor, patchFor, postFor, putFor, routerFor, schemaFor };